+++ src/api/telegram/__init__.py
@@ -2,7 +2,8 @@ from .methods import (
     send_message,
     get_file,
     send_chat_action,
-    edit_message_text
+    edit_message_text,
+    send_photo
 )
 from .requests import (
     create_file_download_url