+++ src/api/yandex/__init__.py
@@ -2,7 +2,8 @@ from .methods import (
     get_access_token,
     upload_file_with_url,
     create_folder,
-    publish
+    publish,
+    get_disk_info
 )
 from .requests import (
     make_link_request,