+++ src/api/yandex/methods.py
@@ -53,3 +53,17 @@ def publish(user_token: str, **kwargs):
         data=kwargs,
         user_token=user_token
     )
+
+
+def get_disk_info(user_token: str, **kwargs):
+    """
+    https://yandex.ru/dev/disk/api/reference/capacity-docpage/
+
+    - see `api/request.py` documentation for more.
+    """
+    return make_disk_request(
+        http_method="GET",
+        api_method="",
+        data=kwargs,
+        user_token=user_token
+    )