+++ src/blueprints/telegram_bot/webhook/views.py
@@ -74,7 +74,8 @@ def route_command(command: str) -> None:
         CommandNames.UPLOAD_VOICE.value: commands.upload_voice_handler,
         CommandNames.UPLOAD_URL.value: commands.upload_url_handler,
         CommandNames.CREATE_FOLDER.value: commands.create_folder_handler,
-        CommandNames.PUBLISH.value: commands.publish_handler
+        CommandNames.PUBLISH.value: commands.publish_handler,
+        CommandNames.SPACE.value: commands.space_handler
     }
     method = routes.get(command, commands.unknown_handler)